fix cors progress
