--- src/vs/workbench/contrib/void/browser/misc/oldpackage.json
@@ -1,168 +1,0 @@
-{
-  "name": "void",
-  "publisher": "void",
-  "displayName": "Void",
-  "description": "",
-  "version": "0.0.1",
-  "engines": {
-    "vscode": "*"
-  },
-  "categories": [
-    "Other"
-  ],
-  "enabledApiProposals": [
-    "editorInsets"
-  ],
-  "activationEvents": [],
-  "main": "./out/extension/extension.js",
-  "contributes": {
-    "configuration": {
-      "title": "Void",
-      "properties": {}
-    },
-    "commands": [
-      {
-        "command": "typeInspector.inspect",
-        "title": "Inspect Types of All Variables"
-      },
-      {
-        "command": "void.ctrl+l",
-        "title": "Show Sidebar"
-      },
-      {
-        "command": "void.ctrl+k",
-        "title": "Make Inline Edit"
-      },
-      {
-        "command": "void.acceptDiff",
-        "title": "Approve Diff"
-      },
-      {
-        "command": "void.rejectDiff",
-        "title": "Discard Diff"
-      },
-      {
-        "command": "void.startNewThread",
-        "title": "Start a new chat",
-        "icon": "$(add)"
-      },
-      {
-        "command": "void.toggleThreadSelector",
-        "title": "View past chats",
-        "icon": "$(history)"
-      },
-      {
-        "command": "void.toggleSettings",
-        "title": "Void settings",
-        "icon": "$(settings-gear)"
-      }
-    ],
-    "viewsContainers": {
-      "activitybar": [
-        {
-          "id": "voidViewContainer",
-          "title": "Chat",
-          "icon": "$(hubot)"
-        }
-      ]
-    },
-    "views": {
-      "voidViewContainer": [
-        {
-          "type": "webview",
-          "id": "void.viewnumberone",
-          "name": "Void"
-        }
-      ]
-    },
-    "keybindings": [
-      {
-        "command": "void.ctrl+l",
-        "key": "ctrl+l",
-        "mac": "cmd+l"
-      },
-      {
-        "command": "void.ctrl+k",
-        "key": "ctrl+k",
-        "mac": "cmd+k"
-      }
-    ],
-    "menus": {
-      "view/title": [
-        {
-          "command": "void.startNewThread",
-          "when": "view == 'void.viewnumberone'",
-          "group": "navigation"
-        },
-        {
-          "command": "void.toggleThreadSelector",
-          "when": "view == 'void.viewnumberone'",
-          "group": "navigation"
-        },
-        {
-          "command": "void.toggleSettings",
-          "when": "view == 'void.viewnumberone'",
-          "group": "navigation"
-        }
-      ]
-    }
-  },
-  "scripts": {
-    "vscode:prepublish": "npm run compile",
-    "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./",
-    "build": "rimraf dist && node build/build.js",
-    "pretest": "tsc -p ./ && eslint src --ext ts",
-    "test": "vscode-test"
-  },
-  "devDependencies": {
-    "@anthropic-ai/sdk": "^0.31.0",
-    "@eslint/js": "^9.9.1",
-    "@google/generative-ai": "^0.21.0",
-    "@monaco-editor/react": "^4.6.0",
-    "@rrweb/types": "^2.0.0-alpha.17",
-    "@types/diff": "^5.2.2",
-    "@types/diff-match-patch": "^1.0.36",
-    "@types/jest": "^29.5.12",
-    "@types/lodash": "^4.17.12",
-    "@types/mocha": "^10.0.8",
-    "@types/node": "^22.5.1",
-    "@types/react": "^18.3.4",
-    "@types/react-dom": "^18.3.0",
-    "@types/react-syntax-highlighter": "^15.5.13",
-    "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.3.0",
-    "@typescript-eslint/parser": "^8.3.0",
-    "@vscode/test-cli": "^0.0.10",
-    "@vscode/test-electron": "2.4.1",
-    "autoprefixer": "^10.4.20",
-    "diff-match-patch": "^1.0.5",
-    "esbuild": "^0.23.1",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.35.1",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "globals": "^15.9.0",
-    "lodash": "^4.17.21",
-    "marked": "^14.1.0",
-    "ollama": "^0.5.9",
-    "openai": "^4.70.2",
-    "postcss": "^8.4.41",
-    "posthog-js": "^1.176.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-markdown": "^9.0.1",
-    "react-syntax-highlighter": "^15.6.1",
-    "rimraf": "^6.0.1",
-    "rrweb-snapshot": "^2.0.0-alpha.4",
-    "tailwindcss": "^3.4.10",
-    "typescript": "5.5.4",
-    "typescript-eslint": "^8.3.0",
-    "uuid": "^10.0.0"
-  },
-  "dependencies": {
-    "lru-cache": "^11.0.2",
-    "tree-sitter": "^0.21.1",
-    "tree-sitter-javascript": "^0.23.1",
-    "tree-sitter-python": "^0.23.4"
-  }
-}
