+++ src/vs/workbench/services/void/electron-sandbox/sendLLMMessage.ts
@@ -1,0 +1,17 @@
+import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
+import { ISendLLMMessageService } from '../common/sendLLMMessage.js';
+
+
+// NODE IMPLEMENTATION OF SENDLLMMESSAGE
+
+export class SendLLMMessageService implements ISendLLMMessageService {
+	readonly _serviceBrand: undefined;
+
+	async sendLLMMessage(data: any): Promise<any> {
+		console.log('NODE sendLLMMessage', data);
+		// Your existing logic to send a message to the server
+		// For example:
+		// return fetch('https://your-server.com/api', { method: 'POST', body: JSON.stringify(data) });
+	}
+}
+registerSingleton(ISendLLMMessageService, SendLLMMessageService, InstantiationType.Delayed);
