+++ src/vs/workbench/workbench.web.main.internal.ts
@@ -32,6 +32,14 @@ import './browser/web.main.js';
 //#endregion
 
 
+
+//#region --- Void
+// Void added this (modeling off of import '.*clipboardservice.js'):
+import './services/void/browser/sendLLMMessage.js';
+//#endregion
+
+
+
 //#region --- workbench services
 
 import './services/integrity/browser/integrityService.js';
